Fix bug when clicking on item
Summary:
Fixes a bug intruduced by D128, that caused move commands to be added
when clicking on an item without moving it. This now check if at least one of
the items has to move.

Test Plan:
Clicking on an item without moving it no longer results in a move being added
to the undo stack.

Reviewers: Joel

Reviewed By: Joel

Differential Revision: https://phabricator.simulton.com/D130