Keep wires connected when they are moved
Summary:
When a wire is moved, keep it connected to the connectors it is connected to
and keep the junction on their wire.

Resolves T1522

Test Plan:
Create a wire between two nodes and a wire between a third node and the first
wire. Using the rubber-band selection, any set of items can be moved without a
wire disconnecting from anything.

Reviewers: Joel

Reviewed By: Joel

Maniphest Tasks: T1522

Differential Revision: https://phabricator.simulton.com/D165