Move selected wire to a higher z-index
Summary:
This makes it possible to move a point that is underneath another wire. This
might not be the best way to do it because it depends on what z-index the wire
currently is on but for now it's better than nothing.

Test Plan:
When two wires overlap eachother, whichever one is selected is moved to the
top.

Reviewers: Joel

Reviewed By: Joel

Differential Revision: https://phabricator.simulton.com/D169