Fix labels snapping when loading a file
Summary:
I'm not entierly sure why all the items are snapped to the grid when the
settings are set, but I decided to leave it as is for all the items that should
snap to the grid.
Resolves T1565

Test Plan: Opening a file doesn't snap all the labels to the grid.

Reviewers: Joel

Reviewed By: Joel

Maniphest Tasks: T1565

Differential Revision: https://phabricator.simulton.com/D172