Adapt wire ending behavior
Summary:
Modifies the behaviour when the mouse is pressed so that when both ends of the
current wire are connected to either a connector or another wire it is
finished.
Resolves T1577

Test Plan: A single click on a wire or connector terminates the wire and stays in wire mode.

Reviewers: Joel

Reviewed By: Joel

Maniphest Tasks: T1577

Differential Revision: https://phabricator.simulton.com/D181