Add a signal to the scene when the mouse hovers over a connector
Summary:
`Scene::connectorHighlighted()` is emitted when the mouse enters a connector
and when it leaves it. When it enters the parameter is a `shared_ptr` to the
connector and when the mouse leaves the connector it's `nullptr`.

Resolves T1594

Test Plan: The signal is emitted when the mouse enters and leaves a connector.

Reviewers: Joel

Reviewed By: Joel

Maniphest Tasks: T1594

Differential Revision: https://phabricator.simulton.com/D190