demo: improve FancyWire connection indication rendering
Previously, we simply checked whether a wire point is on top of a connector. However, this can lead to cases where it looks like the connector is connected without it actually being connected. This is now improved by directly querying the connection state of a wire point from the wire manager system.